+++ tests/dialect_test.py
@@ -88,7 +88,7 @@ class SqlAlchemyDialectTest(TestCase):
         )
         self.fake_cursor.fetchall = MagicMock(return_value=[["id"], ["id2"], ["id3"]])
 
-        eq_(insp.get_pk_constraint("characters")['constrained_columns'], {"id", "id2", "id3"})
+        eq_(insp.get_pk_constraint("characters")['constrained_columns'], ["id", "id2", "id3"])
         self.fake_cursor.fetchall.assert_called_once_with()
         in_("information_schema.key_column_usage", self.executed_statement)
         in_("table_catalog = ?", self.executed_statement)
@@ -103,7 +103,7 @@ class SqlAlchemyDialectTest(TestCase):
         )
         self.fake_cursor.fetchall = MagicMock(return_value=[["id"], ["id2"], ["id3"]])
 
-        eq_(insp.get_pk_constraint("characters")['constrained_columns'], {"id", "id2", "id3"})
+        eq_(insp.get_pk_constraint("characters")['constrained_columns'], ["id", "id2", "id3"])
         self.fake_cursor.fetchall.assert_called_once_with()
         in_("information_schema.key_column_usage", self.executed_statement)
         in_("table_schema = ?", self.executed_statement)