+++ CHANGES.md
@@ -6,6 +6,9 @@
 - Added `CrateIdentifierPreparer`, in order to quote reserved words
   like `object` properly, for example when used as column names.
 - Fixed `CrateDialect.get_pk_constraint` to return `list` instead of `set` type
+- Added re-usable patches and polyfills from application adapters.
+  New utilities: `patch_autoincrement_timestamp`, `refresh_after_dml`,
+  `check_uniqueness_factory`
 
 ## 2024/06/13 0.37.0
 - Added support for CrateDB's [FLOAT_VECTOR] data type and its accompanying