+++ CHANGES.md
@@ -11,6 +11,8 @@
   `check_uniqueness_factory`
 - Added `table_kwargs` context manager to enable pandas/Dask to support
   CrateDB dialect table options.
+- Fixed SQL rendering of special DDL table options in `CrateDDLCompiler`.
+  Before, configuring `crate_"translog.durability"` was not possible.
 
 ## 2024/06/13 0.37.0
 - Added support for CrateDB's [FLOAT_VECTOR] data type and its accompanying